MPSDK Sample App - Refactored sample app to split out screens and functionality
